Aggregator response file rename
