+++ shared/aggregator-responses/balance.response.dto.ts
