+++ shared/aggregator-responses/sum.response.dto.ts
